Move automatic registered Nova Resources to Resources subfolder
