+++ app/Nova/Resources/Subscription.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Nova;
+namespace App\Nova\Resources;
 
 use App\Domain\Subscriptions\BillingInterval;
 use App\Domain\Subscriptions\Currency;
