+++ app/Nova/Resources/User.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Nova;
+namespace App\Nova\Resources;
 
 use App\Domain\Users\Models\UserModel;
 use Illuminate\Validation\Rules;
